+++ application/views/components/booking_footer.php
@@ -3,6 +3,8 @@
  * Local variables.
  *
  * @var bool $display_login_button
+ * @var string $legal_notice_url
+ * @var string $imprint_url
  */
 ?>
 
@@ -11,6 +13,16 @@
         <span class="footer-powered-by small d-block w-100 w-md-50 text-center text-md-start p-1 pe-md-0">
             Powered By
             <a href="https://easyappointments.org" target="_blank">Easy!Appointments</a>
+
+            <?php if (!empty($legal_notice_url)): ?>
+                <span class="mx-1">|</span>
+                <a href="<?= e($legal_notice_url) ?>" target="_blank"><?= lang('legal_notice') ?></a>
+            <?php endif; ?>
+
+            <?php if (!empty($imprint_url)): ?>
+                <span class="mx-1">|</span>
+                <a href="<?= e($imprint_url) ?>" target="_blank"><?= lang('imprint') ?></a>
+            <?php endif; ?>
         </span>
 
         <span class="footer-options d-block w-100 w-md-50 text-center text-md-end">